+++ library/Collections/BwdTrans.cpp
@@ -35,6 +35,7 @@
 #include <Collections/CoalescedGeomData.h>
 #include <Collections/MatrixFreeBase.h>
 #include <Collections/Operator.h>
+
 #include <MatrixFreeOps/Operator.hpp>
 #include <boost/core/ignore_unused.hpp>
 
@@ -52,9 +53,28 @@ using LibUtilities::eTetrahedron;
 using LibUtilities::eTriangle;
 
 /**
+ * @brief Backward transform help class to calculate the size of the collection
+ * that is given as an input and as an output to the BwdTrans Operator. The size
+ * evaluation takes into account the conversion from the coefficient space to
+ * the physical space
+ */
+class BwdTrans_Helper : virtual public Operator
+{
+protected:
+    BwdTrans_Helper()
+    {
+        // expect input to be number of elements by the number of coefficients
+        m_inputSize = m_numElmt * m_stdExp->GetNcoeffs();
+        // expect input to be number of elements by the number of quad points
+        m_outputSize = m_numElmt * m_stdExp->GetTotPoints();
+    }
+};
+
+/**
  * @brief Backward transform operator using standard matrix approach.
  */
-class BwdTrans_StdMat final : public Operator
+class BwdTrans_StdMat final : virtual public Operator,
+                              virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_StdMat)
@@ -96,7 +116,7 @@ private:
     BwdTrans_StdMat(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                     CoalescedGeomDataSharedPtr pGeomData,
                     StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper()
     {
         StdRegions::StdMatrixKey key(StdRegions::eBwdTrans,
                                      m_stdExp->DetShapeType(), *m_stdExp);
@@ -143,7 +163,9 @@ OperatorKey BwdTrans_StdMat::m_typeArr[] = {
 /**
  * @brief Backward transform operator using matrix free operators.
  */
-class BwdTrans_MatrixFree final : public Operator, MatrixFreeOneInOneOut
+class BwdTrans_MatrixFree final : virtual public Operator,
+                                  MatrixFreeOneInOneOut,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_MatrixFree)
@@ -195,7 +217,7 @@ private:
     BwdTrans_MatrixFree(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           MatrixFreeOneInOneOut(pCollExp[0]->GetStdExp()->GetNcoeffs(),
                                 pCollExp[0]->GetStdExp()->GetTotPoints(),
                                 pCollExp.size())
@@ -249,7 +271,8 @@ OperatorKey BwdTrans_MatrixFree::m_typeArr[] = {
 /**
  * @brief Backward transform operator using default StdRegions operator
  */
-class BwdTrans_IterPerExp final : public Operator
+class BwdTrans_IterPerExp final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_IterPerExp)
@@ -293,7 +316,7 @@ private:
     BwdTrans_IterPerExp(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper()
     {
     }
 };
@@ -335,7 +358,8 @@ OperatorKey BwdTrans_IterPerExp::m_typeArr[] = {
 /**
  * @brief Backward transform operator using LocalRegions implementation.
  */
-class BwdTrans_NoCollection final : public Operator
+class BwdTrans_NoCollection final : virtual public Operator,
+                                    virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_NoCollection)
@@ -383,7 +407,7 @@ private:
     BwdTrans_NoCollection(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                           CoalescedGeomDataSharedPtr pGeomData,
                           StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper()
     {
         m_expList = pCollExp;
     }
@@ -426,7 +450,8 @@ OperatorKey BwdTrans_NoCollection::m_typeArr[] = {
 /**
  * @brief Backward transform operator using sum-factorisation (Segment)
  */
-class BwdTrans_SumFac_Seg final : public Operator
+class BwdTrans_SumFac_Seg final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Seg)
@@ -479,7 +504,7 @@ private:
     BwdTrans_SumFac_Seg(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nmodes0(m_stdExp->GetBasisNumModes(0)),
           m_colldir0(m_stdExp->GetBasis(0)->Collocation()),
@@ -498,7 +523,8 @@ OperatorKey BwdTrans_SumFac_Seg::m_type =
 /**
  * @brief Backward transform operator using sum-factorisation (Quad)
  */
-class BwdTrans_SumFac_Quad final : public Operator
+class BwdTrans_SumFac_Quad final : virtual public Operator,
+                                   virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Quad)
@@ -584,7 +610,7 @@ private:
     BwdTrans_SumFac_Quad(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                          CoalescedGeomDataSharedPtr pGeomData,
                          StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nmodes0(m_stdExp->GetBasisNumModes(0)),
@@ -607,7 +633,8 @@ OperatorKey BwdTrans_SumFac_Quad::m_type =
 /**
  * @brief Backward transform operator using sum-factorisation (Tri)
  */
-class BwdTrans_SumFac_Tri final : public Operator
+class BwdTrans_SumFac_Tri final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Tri)
@@ -708,7 +735,8 @@ OperatorKey BwdTrans_SumFac_Tri::m_type =
         BwdTrans_SumFac_Tri::create, "BwdTrans_SumFac_Tri");
 
 /// Backward transform operator using sum-factorisation (Hex)
-class BwdTrans_SumFac_Hex final : public Operator
+class BwdTrans_SumFac_Hex final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Hex)
@@ -794,7 +822,7 @@ private:
     BwdTrans_SumFac_Hex(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(pCollExp[0]->GetNumPoints(0)),
           m_nquad1(pCollExp[0]->GetNumPoints(1)),
           m_nquad2(pCollExp[0]->GetNumPoints(2)),
@@ -822,7 +850,8 @@ OperatorKey BwdTrans_SumFac_Hex::m_type =
 /**
  * @brief Backward transform operator using sum-factorisation (Tet)
  */
-class BwdTrans_SumFac_Tet final : public Operator
+class BwdTrans_SumFac_Tet final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Tet)
@@ -962,7 +991,7 @@ private:
     BwdTrans_SumFac_Tet(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),
@@ -997,7 +1026,8 @@ OperatorKey BwdTrans_SumFac_Tet::m_type =
 /**
  * @brief Backward transform operator using sum-factorisation (Prism)
  */
-class BwdTrans_SumFac_Prism final : public Operator
+class BwdTrans_SumFac_Prism final : virtual public Operator,
+                                    virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Prism)
@@ -1107,7 +1137,7 @@ private:
     BwdTrans_SumFac_Prism(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                           CoalescedGeomDataSharedPtr pGeomData,
                           StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),
@@ -1141,7 +1171,8 @@ OperatorKey BwdTrans_SumFac_Prism::m_type =
 /**
  * @brief Backward transform operator using sum-factorisation (Pyr)
  */
-class BwdTrans_SumFac_Pyr final : public Operator
+class BwdTrans_SumFac_Pyr final : virtual public Operator,
+                                  virtual public BwdTrans_Helper
 {
 public:
     OPERATOR_CREATE(BwdTrans_SumFac_Pyr)
@@ -1271,7 +1302,7 @@ private:
     BwdTrans_SumFac_Pyr(vector<StdRegions::StdExpansionSharedPtr> pCollExp,
                         CoalescedGeomDataSharedPtr pGeomData,
                         StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), BwdTrans_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),