+++ library/Collections/IProductWRTDerivBase.cpp
@@ -32,14 +32,13 @@
 //
 ///////////////////////////////////////////////////////////////////////////////
 
-#include <boost/core/ignore_unused.hpp>
-
-#include <MatrixFreeOps/Operator.hpp>
-
 #include <Collections/Collection.h>
 #include <Collections/IProduct.h>
 #include <Collections/MatrixFreeBase.h>
 #include <Collections/Operator.h>
+
+#include <MatrixFreeOps/Operator.hpp>
+#include <boost/core/ignore_unused.hpp>
 
 using namespace std;
 
@@ -55,9 +54,28 @@ using LibUtilities::eTetrahedron;
 using LibUtilities::eTriangle;
 
 /**
+ * @brief Inner product deriv base help class to calculate the size of the
+ * collection that is given as an input and as an output to the
+ * IProductWRTDerivBase Operator. The size evaluation takes into account the
+ * conversion from the physical space to the coefficient space.
+ */
+class IProductWRTDerivBase_Helper : virtual public Operator
+{
+protected:
+    IProductWRTDerivBase_Helper()
+    {
+        // expect input to be number of elements by the number of quad points
+        m_inputSize = m_numElmt * m_stdExp->GetTotPoints();
+        // expect input to be number of elements by the number of coefficients
+        m_outputSize = m_numElmt * m_stdExp->GetNcoeffs();
+    }
+};
+
+/**
  * @brief Inner product WRT deriv base operator using standard matrix approach
  */
-class IProductWRTDerivBase_StdMat final : public Operator
+class IProductWRTDerivBase_StdMat final : virtual public Operator,
+                                          IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_StdMat)
@@ -70,7 +88,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         int nPhys  = m_stdExp->GetTotPoints();
         int ntot   = m_numElmt * nPhys;
         int nmodes = m_stdExp->GetNcoeffs();
@@ -186,7 +203,7 @@ private:
     IProductWRTDerivBase_StdMat(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper()
     {
         m_dim     = pCollExp[0]->GetShapeDimension();
         m_coordim = pCollExp[0]->GetCoordim();
@@ -262,8 +279,9 @@ OperatorKey IProductWRTDerivBase_StdMat::m_typeArr[] = {
 /**
  * @brief Inner product operator using operator using matrix free operators.
  */
-class IProductWRTDerivBase_MatrixFree final : public Operator,
-                                              MatrixFreeMultiInOneOut
+class IProductWRTDerivBase_MatrixFree final : virtual public Operator,
+                                              MatrixFreeMultiInOneOut,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_MatrixFree)
@@ -366,7 +384,8 @@ private:
           MatrixFreeMultiInOneOut(pCollExp[0]->GetCoordim(),
                                   pCollExp[0]->GetStdExp()->GetTotPoints(),
                                   pCollExp[0]->GetStdExp()->GetNcoeffs(),
-                                  pCollExp.size())
+                                  pCollExp.size()),
+          IProductWRTDerivBase_Helper()
     {
         // Check if deformed
         const auto dim = pCollExp[0]->GetStdExp()->GetShapeDimension();
@@ -433,7 +452,8 @@ OperatorKey IProductWRTDerivBase_MatrixFree::m_typeArr[] = {
 /**
  * @brief Inner product WRT deriv base operator using element-wise operation
  */
-class IProductWRTDerivBase_IterPerExp final : public Operator
+class IProductWRTDerivBase_IterPerExp final : virtual public Operator,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_IterPerExp)
@@ -446,7 +466,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -564,7 +583,7 @@ private:
     IProductWRTDerivBase_IterPerExp(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper()
     {
         m_dim     = pCollExp[0]->GetShapeDimension();
         m_coordim = pCollExp[0]->GetCoordim();
@@ -624,7 +643,8 @@ OperatorKey IProductWRTDerivBase_IterPerExp::m_typeArr[] = {
  * @brief Inner product WRT deriv base operator using LocalRegions
  * implementation.
  */
-class IProductWRTDerivBase_NoCollection final : public Operator
+class IProductWRTDerivBase_NoCollection final : virtual public Operator,
+                                                IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_NoCollection)
@@ -693,7 +713,7 @@ private:
     IProductWRTDerivBase_NoCollection(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors)
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper()
     {
         m_expList = pCollExp;
         m_dim     = pCollExp[0]->GetNumBases();
@@ -749,7 +769,8 @@ OperatorKey IProductWRTDerivBase_NoCollection::m_typeArr[] = {
  * @brief Inner product WRT deriv base operator using sum-factorisation
  * (Segment)
  */
-class IProductWRTDerivBase_SumFac_Seg final : public Operator
+class IProductWRTDerivBase_SumFac_Seg final : virtual public Operator,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Seg)
@@ -836,7 +857,7 @@ private:
     IProductWRTDerivBase_SumFac_Seg(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nmodes0(m_stdExp->GetBasisNumModes(0)),
           m_derbase0(m_stdExp->GetBasis(0)->GetDbdata())
@@ -858,7 +879,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Seg::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Quad)
  */
-class IProductWRTDerivBase_SumFac_Quad final : public Operator
+class IProductWRTDerivBase_SumFac_Quad final : virtual public Operator,
+                                               IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Quad)
@@ -871,7 +893,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -969,7 +990,7 @@ private:
     IProductWRTDerivBase_SumFac_Quad(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nmodes0(m_stdExp->GetBasisNumModes(0)),
@@ -999,7 +1020,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Quad::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Tri)
  */
-class IProductWRTDerivBase_SumFac_Tri final : public Operator
+class IProductWRTDerivBase_SumFac_Tri final : virtual public Operator,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Tri)
@@ -1043,7 +1065,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -1157,7 +1178,7 @@ private:
     IProductWRTDerivBase_SumFac_Tri(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nmodes0(m_stdExp->GetBasisNumModes(0)),
@@ -1219,7 +1240,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Tri::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Hex)
  */
-class IProductWRTDerivBase_SumFac_Hex final : public Operator
+class IProductWRTDerivBase_SumFac_Hex final : virtual public Operator,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Hex)
@@ -1232,7 +1254,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -1344,7 +1365,7 @@ private:
     IProductWRTDerivBase_SumFac_Hex(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),
@@ -1380,7 +1401,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Hex::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Tet)
  */
-class IProductWRTDerivBase_SumFac_Tet : public Operator
+class IProductWRTDerivBase_SumFac_Tet : virtual public Operator,
+                                        IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Tet)
@@ -1444,7 +1466,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -1581,7 +1602,7 @@ private:
     IProductWRTDerivBase_SumFac_Tet(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),
@@ -1617,7 +1638,6 @@ private:
             {
                 for (int k = 0; k < m_nquad2; ++k)
                 {
-
                     m_fac0[i + j * m_nquad0 + k * m_nquad0 * m_nquad1] =
                         4.0 / ((1 - z1[j]) * (1 - z2[k]));
                     m_fac1[i + j * m_nquad0 + k * m_nquad0 * m_nquad1] =
@@ -1651,7 +1671,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Tet::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Prism)
  */
-class IProductWRTDerivBase_SumFac_Prism final : public Operator
+class IProductWRTDerivBase_SumFac_Prism final : virtual public Operator,
+                                                IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Prism)
@@ -1710,7 +1731,6 @@ public:
                     Array<OneD, NekDouble> &entry3,
                     Array<OneD, NekDouble> &wsp) final
     {
-
         unsigned int nPhys  = m_stdExp->GetTotPoints();
         unsigned int ntot   = m_numElmt * nPhys;
         unsigned int nmodes = m_stdExp->GetNcoeffs();
@@ -1832,7 +1852,7 @@ private:
     IProductWRTDerivBase_SumFac_Prism(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),
@@ -1896,7 +1916,8 @@ OperatorKey IProductWRTDerivBase_SumFac_Prism::m_type =
 /**
  * @brief Inner product WRT deriv base operator using sum-factorisation (Pyr)
  */
-class IProductWRTDerivBase_SumFac_Pyr final : public Operator
+class IProductWRTDerivBase_SumFac_Pyr final : virtual public Operator,
+                                              IProductWRTDerivBase_Helper
 {
 public:
     OPERATOR_CREATE(IProductWRTDerivBase_SumFac_Pyr)
@@ -2093,7 +2114,7 @@ private:
     IProductWRTDerivBase_SumFac_Pyr(
         vector<StdRegions::StdExpansionSharedPtr> pCollExp,
         CoalescedGeomDataSharedPtr pGeomData, StdRegions::FactorMap factors)
-        : Operator(pCollExp, pGeomData, factors),
+        : Operator(pCollExp, pGeomData, factors), IProductWRTDerivBase_Helper(),
           m_nquad0(m_stdExp->GetNumPoints(0)),
           m_nquad1(m_stdExp->GetNumPoints(1)),
           m_nquad2(m_stdExp->GetNumPoints(2)),